+++ src/components/desktop-menu/DesktopMenu.jsx
@@ -6,13 +6,11 @@ const DesktopMenu = ({ links, loggedIn }) => (
 	<ul id="desktop-menu">
 		{links
 			.filter((x) => !x.hideOnDesktop)
-			.map((x) => {
-				x.onClick = null;
-				x.forMobile = null;
-				return x;
-			})
+
 			.filter((x) => x.hideOptions({ loggedIn }))
-			.map(MenuItem)}
+			.map((link, i) => {
+				return <MenuItem {...link} i={i} key={i} />;
+			})}
 	</ul>
 );