+++ src/components/menu-item/MenuItem.jsx
@@ -1,12 +1,24 @@
+import Link from 'next/link';
 import React from 'react';
 
-const MenuItem = ({ text, url, target, rel, onClick, forMobile }, i) => {
-	if (url) {
+const MenuItem = ({
+	text,
+	url,
+	target,
+	rel,
+	onClick,
+	isMobile,
+	toggleMenu,
+	i,
+}) => {
+	if (url && !onClick) {
 		return (
-			<li key={i} onClick={forMobile ? onClick : null}>
-				<a href={url} target={target} rel={rel}>
-					{text}
-				</a>
+			<li key={i} onClick={isMobile ? toggleMenu : null}>
+				<Link href={url}>
+					<a target={target} rel={rel}>
+						{text}
+					</a>
+				</Link>
 			</li>
 		);
 	} else {